// src/lib/api/headscale.d.ts
export interface paths {
    "/api/v1/apikey": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_ListApiKeys"];
        put?: never;
        /** --- ApiKeys start --- */
        post: operations["HeadscaleService_CreateApiKey"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/apikey/expire": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_ExpireApiKey"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v1/node": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_ListNodes"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/api/v1/node/register": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_RegisterNode"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_GetNode"];
        put?: never;
        post?: never;
        delete: operations["HeadscaleService_DeleteNode"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}/expire": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_ExpireNode"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}/rename/{newName}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_RenameNode"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}/routes": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_GetNodeRoutes"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}/tags": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_SetTags"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/node/{nodeId}/user": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_MoveNode"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/policy": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** --- Policy start --- */
        get: operations["HeadscaleService_GetPolicy"];
        put: operations["HeadscaleService_SetPolicy"];
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/preauthkey": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_ListPreAuthKeys"];
        put?: never;
        /** --- PreAuthKeys start --- */
        post: operations["HeadscaleService_CreatePreAuthKey"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/preauthkey/expire": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_ExpirePreAuthKey"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/routes": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** --- Route start --- */
        get: operations["HeadscaleService_GetRoutes"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/routes/{routeId}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post?: never;
        delete: operations["HeadscaleService_DeleteRoute"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/routes/{routeId}/disable": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_DisableRoute"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/routes/{routeId}/enable": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_EnableRoute"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/user": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["HeadscaleService_ListUsers"];
        put?: never;
        post: operations["HeadscaleService_CreateUser"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/user/{name}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** --- User start --- */
        get: operations["HeadscaleService_GetUser"];
        put?: never;
        post?: never;
        delete: operations["HeadscaleService_DeleteUser"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/api/v1/user/{oldName}/rename/{newName}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get?: never;
        put?: never;
        post: operations["HeadscaleService_RenameUser"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        HeadscaleServiceSetTagsBody: {
            tags?: string[];
        };
        protobufAny: {
            "@type"?: string;
        } & {
            [key: string]: Record<string, never>;
        };
        rpcStatus: {
            /** Format: int32 */
            code?: number;
            message?: string;
            details?: components["schemas"]["protobufAny"][];
        };
        v1ApiKey: {
            /** Format: uint64 */
            id?: string;
            prefix?: string;
            /** Format: date-time */
            expiration?: string;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            lastSeen?: string;
        };
        v1BackfillNodeIPsResponse: {
            changes?: string[];
        };
        v1CreateApiKeyRequest: {
            /** Format: date-time */
            expiration?: string;
        };
        v1CreateApiKeyResponse: {
            apiKey?: string;
        };
        v1CreatePreAuthKeyRequest: {
            user?: string;
            reusable?: boolean;
            ephemeral?: boolean;
            /** Format: date-time */
            expiration?: string;
            aclTags?: string[];
        };
// ...
        v1Node: {
            /** Format: uint64 */
            id?: string;
            machineKey?: string;
            nodeKey?: string;
            discoKey?: string;
            ipAddresses?: string[];
            name?: string;
            user?: components["schemas"]["v1User"];
            /** Format: date-time */
            lastSeen?: string;
            /** Format: date-time */
            expiry?: string;
            preAuthKey?: components["schemas"]["v1PreAuthKey"];
            /** Format: date-time */
            createdAt?: string;
            registerMethod?: components["schemas"]["v1RegisterMethod"];
            forcedTags?: string[];
            invalidTags?: string[];
            validTags?: string[];
            givenName?: string;
            online?: boolean;
        };
        v1PreAuthKey: {
            user?: string;
            id?: string;
            key?: string;
            reusable?: boolean;
            ephemeral?: boolean;
            used?: boolean;
            /** Format: date-time */
            expiration?: string;
            /** Format: date-time */
            createdAt?: string;
            aclTags?: string[];
        };
        /**
         * @default REGISTER_METHOD_UNSPECIFIED
         * @enum {string}
         */
        v1RegisterMethod: "REGISTER_METHOD_UNSPECIFIED" | "REGISTER_METHOD_AUTH_KEY" | "REGISTER_METHOD_CLI" | "REGISTER_METHOD_OIDC";
        v1RegisterNodeResponse: {
            node?: components["schemas"]["v1Node"];
        };
        v1RenameNodeResponse: {
            node?: components["schemas"]["v1Node"];
        };
        v1RenameUserResponse: {
            user?: components["schemas"]["v1User"];
        };
        v1Route: {
            /** Format: uint64 */
            id?: string;
            node?: components["schemas"]["v1Node"];
            prefix?: string;
            advertised?: boolean;
            enabled?: boolean;
            isPrimary?: boolean;
            /** Format: date-time */
            createdAt?: string;
            /** Format: date-time */
            updatedAt?: string;
            /** Format: date-time */
            deletedAt?: string;
        };
        v1SetPolicyRequest: {
            policy?: string;
        };
        v1SetPolicyResponse: {
            policy?: string;
            /** Format: date-time */
            updatedAt?: string;
        };
        v1SetTagsResponse: {
            node?: components["schemas"]["v1Node"];
        };
        v1User: {
            id?: string;
            name?: string;
            /** Format: date-time */
            createdAt?: string;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}